Update index.html
Added a CDN link to the latest available and stable version of Bootstrap. The version will be updated with this template whenever possible and necessary. If an older project needs to be updated to the latest, please manually update it in the project. There is no need to get from here (unless for copy/reference).